// src/types.ts
export type UserRole = "FARMER" | "BUYER" | "LOGISTICS" | "GOVT_OFFICIAL";

export interface User {
  id: number;
  name: string;
  phone: string;
  email: string;
  role: UserRole;
  fpoName?: string;
  district: string;
  state: string;
  lat: number;
  lng: number;
  trustScore: number;
  verified: boolean;
  kycStatus: string;
  totalTrades: number;
  ratingCount: number;
  createdAt: string;
}

export type QualityGrade = "Grade A" | "Grade B" | "Grade C";

export interface CropListing {
  id: number;
  farmerId: number;
  cropName: string;
  variety: string;
  quantityQuintals: number;
  qualityGrade: QualityGrade;
  harvestDate: string;
  district: string;
  state: string;
  pincode: string;
  lat: number;
  lng: number;
  isOrganic: boolean;
  expectedPricePerQuintal: number;
  mandiBenchmarkPrice: number;
  aiRecommendedMin: number;
  aiRecommendedMax: number;
  aiRecommendedTarget: number;
  status: "ACTIVE" | "SOLD" | "IN_NEGOTIATION" | "EXPIRED";
  notes?: string;
  imageUrl?: string;
  createdAt: string;
  farmerName?: string;
  farmerTrustScore?: number;
  farmerVerified?: boolean;
}
// ...
export type OrderStatus = "PLACED" | "CONFIRMED" | "BATCH_ASSIGNED" | "IN_TRANSIT" | "DELIVERED" | "COMPLETED";
export type PaymentStatus = "PENDING" | "ESCROW_HELD" | "RELEASED_TO_FARMER";

export interface Order {
  id: number;
  orderNumber: string;
  listingId: number;
  buyerId: number;
  farmerId: number;
  cropName: string;
  quantityOrdered: number;
  pricePerQuintal: number;
  totalProduceAmount: number;
  logisticsFee: number;
  platformFee: number;
  totalAmount: number;
  status: OrderStatus;
  paymentStatus: PaymentStatus;
  paymentRef?: string;
  deliveryAddress: string;
  deliveryPincode: string;
  deliveryOtp: string;
  buyerName?: string;
  farmerName?: string;
  createdAt: string;
  updatedAt: string;
}